1. 修复内容管理中，当附件域名不加协议时，修改文章内容是，图片错误得增加两次路径的问题。 2. 优化可视化编辑插件的资源加载，使之更好支持https 3. 修复模版导入，有的用户导入自己开发的模版（zip模版包）时，导入出错的问题。

+++ src/main/java/com/xnx3/wangmarket/admin/controller/SiteController.java
@@ -690,7 +690,7 @@ public class SiteController extends BaseController {
 			sqlService.save(news);
 			
 			//如果有旧图，删除掉旧的图片
-			if(oldTitlePic.length() > 0 && oldTitlePic.indexOf("http://") == -1){
+			if(oldTitlePic.length() > 0 && (oldTitlePic.indexOf("http://") == -1 && oldTitlePic.indexOf("https://") == -1 && oldTitlePic.indexOf("//") == -1)){
 				AttachmentFile.deleteObject("site/"+site.getId()+"/news/"+oldTitlePic);
 			}
